changed returned value in ft_atoi from int to long

+++ includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: yquaro <yquaro@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/22 12:05:21 by yquaro            #+#    #+#             */
-/*   Updated: 2020/01/17 01:56:15 by yquaro           ###   ########.fr       */
+/*   Updated: 2020/02/09 21:55:28 by yquaro           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -174,7 +174,7 @@ char					*ft_strstr(const char *s1, const char *s2);
 char					*ft_strnstr(const char *s1, const char *s2, size_t n);
 int						ft_strcmp(const char *s1, const char *s2);
 int						ft_strncmp(const char *s1, const char *s2, size_t n);
-int						ft_atoi(const char *str);
+long					ft_atoi(const char *str);
 int						ft_isalpha(int c);
 int						ft_isdigit(int c);
 int						ft_is_digitline(const char *line);
